requirements.txt
Flask==3.0.2
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Mail==0.10.0
Flask-Login==0.6.3
Flask-Moment==1.0.6
flask-babel==4.0.0
WTForms==3.1.2
email_validator==2.1.1
Werkzeug==3.0.1
ipykernel==6.29.3
ipython==8.18.1
Bootstrap-Flask==2.3.3
bs4==0.0.2
beautifulsoup4==4.12.3
lxml==5.1.0
numpy==1.26.4
pandas==2.2.1
PyJWT==2.9.0
psycopg2-binary==2.9.9
selenium==4.18.1
python-dotenv==1.0.1
langdetect==1.0.9
elasticsearch==8.11.1
azure-storage-blob==12.24.0